Fix compilation warnings

+++ test/unit_tests/unit_test_image_function.cpp
@@ -386,7 +386,7 @@ namespace Function_Template
 
         const PenguinV_Image::Image16Bit output = ConvertTo16Bit( input );
 
-        return verifyImage( output, intensity * 256 );
+        return verifyImage( output, static_cast<uint16_t>( intensity * 256 ) );
     }
 
     bool form2_ConvertTo16Bit(ConvertTo16BitForm2 ConvertTo16Bit)
@@ -399,7 +399,7 @@ namespace Function_Template
 
         ConvertTo16Bit( input, output );
 
-        return verifyImage( output, intensity[0] * 256 );
+        return verifyImage( output, static_cast<uint16_t>( intensity[0] * 256 ) );
     }
 
     bool form3_ConvertTo16Bit(ConvertTo16BitForm3 ConvertTo16Bit)
@@ -412,7 +412,7 @@ namespace Function_Template
 
         const PenguinV_Image::Image16Bit output = ConvertTo16Bit( input, roiX, roiY, roiWidth, roiHeight );
 
-        return verifyImage( output, intensity * 256 );
+        return verifyImage( output, static_cast<uint16_t>( intensity * 256 ) );
     }
 
     bool form4_ConvertTo16Bit(ConvertTo16BitForm4 ConvertTo16Bit)
@@ -432,7 +432,7 @@ namespace Function_Template
 
         ConvertTo16Bit( input, roiX[0], roiY[0], output, roiX[1], roiY[1], roiWidth, roiHeight );
 
-        return verifyImage( output, roiX[1], roiY[1], roiWidth, roiHeight, intensity[0] * 256 );
+        return verifyImage( output, roiX[1], roiY[1], roiWidth, roiHeight, static_cast<uint16_t>( intensity[0] * 256 ) );
     }
 
     bool form1_ConvertTo8Bit(ConvertTo8BitForm1 ConvertTo8Bit)
@@ -1127,7 +1127,7 @@ namespace Function_Template
 
         const PenguinV_Image::Image output = Invert( input );
 
-        return verifyImage( output, ~intensity );
+        return verifyImage( output, static_cast<uint8_t>( ~intensity ) );
     }
 
     bool form2_Invert(InvertForm2 Invert)
@@ -1137,7 +1137,7 @@ namespace Function_Template
 
         Invert( input[0], input[1] );
 
-        return verifyImage( input[1], ~intensity[0] );
+        return verifyImage( input[1], static_cast<uint8_t>( ~intensity[0] ) );
     }
 
     bool form3_Invert(InvertForm3 Invert)
@@ -1150,7 +1150,7 @@ namespace Function_Template
 
         const PenguinV_Image::Image output = Invert( input, roiX, roiY, roiWidth, roiHeight );
 
-        return equalSize( output, roiWidth, roiHeight ) && verifyImage( output, ~intensity );
+        return equalSize( output, roiWidth, roiHeight ) && verifyImage( output, static_cast<uint8_t>( ~intensity ) );
     }
 
     bool form4_Invert(InvertForm4 Invert)
@@ -1164,7 +1164,7 @@ namespace Function_Template
 
         Invert( image[0], roiX[0], roiY[0], image[1], roiX[1], roiY[1], roiWidth, roiHeight );
 
-        return verifyImage( image[1], roiX[1], roiY[1], roiWidth, roiHeight, ~intensity[0] );
+        return verifyImage( image[1], roiX[1], roiY[1], roiWidth, roiHeight, static_cast<uint8_t>( ~intensity[0] ) );
     }
 
     bool form1_IsEqual(IsEqualForm1 IsEqual)
